pyproject.toml
[project]
name = "loci"
version = "0.1.0"
description = "A personal memory graph server: raw sources + interpretation graph + per-project view, served to any client."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "ryan-waterloo" }]

# Runtime dependencies. Pinned conservatively — we want reproducibility but not
# to be hostile to consumers. Ranges chosen against the April 2026 ecosystem.
dependencies = [
    # Core HTTP / API
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "websockets>=13",
    "pydantic>=2.9,<3",
    "pydantic-settings>=2.5,<3",

    # Storage
    # sqlite-vec ships a loadable extension we attach at connection time. See
    # `loci/db/connection.py`. The Python wheel bundles the .dylib/.so/.dll.
    "sqlite-vec>=0.1.6",

    # Embeddings (local). sentence-transformers handles model loading + batching.
    # Default model is BAAI/bge-small-en-v1.5 (384-dim) for footprint; configurable.
    "sentence-transformers>=3,<5",

    # Document extraction
    # pymupdf is AGPL — we ship as optional to keep the core MIT-clean. PDFs without
    # it fall through to pypdf. See `loci/ingest/extractors.py`.
    "pypdf>=5,<6",
    "beautifulsoup4>=4.12,<5",
    "markdown-it-py>=3,<4",

    # Graph / numerics
    "numpy>=1.26,<3",
    "scipy>=1.13,<2",

    # LLM. pydantic-ai is provider-neutral (Anthropic, OpenAI, OpenRouter, …).
    # We build per-task Agent instances; each task's model is configured by
    # spec string, e.g. "anthropic:claude-opus-4-7" or "openrouter:google/gemini-2.0-flash-exp".
    # We use the `slim` distribution and only pull the provider extras we ship
    # with — the meta `pydantic-ai` package pulls `huggingface-hub>=1.3.4`,
    # which conflicts with sentence-transformers' `<1` pin.
    "pydantic-ai-slim[anthropic,openai]>=1.87,<2",

    # CLI. cyclopts (>=3) is type-hint-driven, supports Union/Literal/docstring
    # help cleanly. Chosen over typer because typer can't surface Literal/Union
    # in --help and forces a `Annotated[X, typer.Option(...)]` ceremony for
    # things cyclopts derives from the type hint. See research note in commit log.
    "cyclopts>=3,<5",
    "rich>=13,<14",

    # MCP (Model Context Protocol)
    "mcp>=1.2,<2",

    # IDs + utilities
    "ulid-py>=1.1,<2",
    "httpx>=0.27,<1",
    "python-multipart>=0.0.10",
]

[project.optional-dependencies]
# pymupdf4llm is the markdown-out, LLM-targeted layer over PyMuPDF. PyMuPDF is
# AGPL-3 — opt-in. Users who need high-quality PDF text install
# `loci[pdf-quality]`. Without it, we fall back to pypdf (BSD).
pdf-quality = ["pymupdf>=1.24,<2", "pymupdf4llm>=0.0.17"]
# NOTE on marker (datalab-to/marker, the OCR-aware PDF→markdown tool):
# marker-pdf 1.10.x hard-pins `anthropic>=0.46.0,<0.47.0`, which conflicts
# with pydantic-ai-slim's `anthropic>=0.96.0`. We therefore CANNOT bundle it
# as a normal extras-tag. Users who want marker should install it into a
# *separate* virtualenv, OR use `uv pip install marker-pdf --no-deps` and
# manage marker's deps manually. `loci/ingest/extractors.py` detects marker
# via a runtime import; no further config is needed once it's importable.
# See docs/sources.md for the gory details.
# Leiden community detection. Used by the absorb job's community pass; optional
# because igraph + leidenalg pull C deps.
graph = ["python-igraph>=0.11", "leidenalg>=0.10"]
# Anthropic-key-free local LLM path (future). Placeholder.
local-llm = []
# Dev-only.
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "pytest-cov>=5,<6",
    "ruff>=0.6,<1",
    "mypy>=1.11,<2",
    "httpx>=0.27,<1",
]

[project.scripts]
loci = "loci.cli:main"

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/loci"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
# E,F: pyflakes/pycodestyle errors. W: warnings. I: import order.
# B: bugbear. UP: pyupgrade. SIM: flake8-simplify.
# Intentionally narrow — this is infrastructure code, not a style museum.
select = ["E", "F", "W", "I", "B", "UP", "SIM"]
ignore = [
    "E501", # line length handled by formatter
    "B008", # FastAPI dependency injection uses `= Depends(...)` defaults
]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers --strict-config"
markers = [
    "slow: tests that hit the embedding model or LLM (skip by default with -m 'not slow')",
    "integration: end-to-end vignettes",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
exclude = ["tests/fixtures/"]
